feat: 💡C端用户接口【修改信息、上传头像】

+++ src/main/java/run/hxtia/workbd/pojo/po/CourseClazz.java
@@ -26,7 +26,7 @@ public class CourseClazz implements Serializable {
     /**
      * 组织ID
      */
-    private Long orgId;
+    private Short orgId;
 
 }
